+++ src/main/java/com/cburch/logisim/instance/InstanceStateImpl.java
@@ -103,7 +103,7 @@ public class InstanceStateImpl implements InstanceState {
     return circuitState.getProject();
   }
 
-  public long getTickCount() {
+  public int getTickCount() {
     return circuitState.getPropagator().getTickCount();
   }
